Chg: Seperate signIn button and signOut button
